add stock subtract, restore quantity fearture

+++ stock/src/main/java/com/trinhvu/stock/model/Stock.java
@@ -23,7 +23,7 @@ public class Stock extends AbstractAuditEntity{
     private Double highPrice;
     private Double lowPrice;
     private Double volume;
-    private Integer availableQuantity;
+    private Long availableQuantity;
     private boolean isPublished;
     private boolean isFeatured;
     private boolean isVisibleIndividually;
